refactor: 고객 request dto 패키지 변경
