+++ src/main/java/org/prgrms/kdt/domain/customer/request/CustomerCreateRequest.java
@@ -1,17 +1,19 @@
-package org.prgrms.kdt.domain.customer.dto;
+package org.prgrms.kdt.domain.customer.request;
 
+import org.prgrms.kdt.domain.customer.model.Customer;
 import org.prgrms.kdt.domain.customer.model.CustomerType;
 
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
+import java.time.LocalDateTime;
+import java.util.UUID;
 
-public class CustomerUpdateRequest {
-
+public class CustomerCreateRequest {
     @NotNull
     private CustomerType customerType;
-    @NotBlank(message = "이름에 공백이 들어갈 수 없습니다.")
+    @NotBlank(message = "이름은 필수입력입니다.")
     @Size(max = 10, message = "이름은 10자 이하로 입력해주세요.")
     private String name;
     @NotBlank(message = "이메일은 필수입력입니다.")
@@ -19,7 +21,7 @@ public class CustomerUpdateRequest {
     @Size(max = 50, message = "이메일은 50자 이하로 입력해주세요.")
     private String email;
 
-    public CustomerUpdateRequest() {
+    public CustomerCreateRequest() {
     }
 
     public CustomerType getCustomerType() {
@@ -46,4 +48,11 @@ public class CustomerUpdateRequest {
         this.email = email;
     }
 
+    public Customer toEntity() {
+        return new Customer(UUID.randomUUID(),
+                name,
+                email,
+                LocalDateTime.now(),
+                LocalDateTime.now());
+    }
 }
