+++ src/main/java/org/prgrms/kdt/domain/customer/request/CustomerUpdateRequest.java
@@ -1,19 +1,17 @@
-package org.prgrms.kdt.domain.customer.dto;
+package org.prgrms.kdt.domain.customer.request;
 
-import org.prgrms.kdt.domain.customer.model.Customer;
 import org.prgrms.kdt.domain.customer.model.CustomerType;
 
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
-import java.time.LocalDateTime;
-import java.util.UUID;
 
-public class CustomerCreateRequest {
+public class CustomerUpdateRequest {
+
     @NotNull
     private CustomerType customerType;
-    @NotBlank(message = "이름은 필수입력입니다.")
+    @NotBlank(message = "이름에 공백이 들어갈 수 없습니다.")
     @Size(max = 10, message = "이름은 10자 이하로 입력해주세요.")
     private String name;
     @NotBlank(message = "이메일은 필수입력입니다.")
@@ -21,7 +19,7 @@ public class CustomerCreateRequest {
     @Size(max = 50, message = "이메일은 50자 이하로 입력해주세요.")
     private String email;
 
-    public CustomerCreateRequest() {
+    public CustomerUpdateRequest() {
     }
 
     public CustomerType getCustomerType() {
@@ -48,11 +46,4 @@ public class CustomerCreateRequest {
         this.email = email;
     }
 
-    public Customer toEntity() {
-        return new Customer(UUID.randomUUID(),
-                name,
-                email,
-                LocalDateTime.now(),
-                LocalDateTime.now());
-    }
 }
