+++ src/test/java/org/prgrms/kdt/domain/customer/service/CustomerServiceTest.java
@@ -67,6 +67,7 @@ class CustomerServiceTest {
 //        UUID savedId = customerService.saveCustomer("park", "a@gmail.com", CustomerType.NORMAL);
         //then
 //        assertThat(savedId).isEqualTo(customerId);
+
     }
